fix: mypy fixes
